requirements.txt
Flask==2.3.3
google-genai==1.14.0
python-dotenv==1.0.0
python-weather==2.0.7
requests==2.31.0 